backend with ns timestamp support

+++ backend/src/main/java/org/example/database/entity/Imu.java
@@ -19,7 +19,7 @@ public class Imu {
     private Long id;
 
     @Column(name = "timestamp_utc")
-    private int timestampUtc;
+    private Long timestampUtc;
 
     @Column(name = "timestamp_ms")
     private int timestampMs;
@@ -45,11 +45,11 @@ public class Imu {
         this.id = id;
     }
 
-    public int getTimestampUtc() {
+    public Long getTimestampUtc() {
         return timestampUtc;
     }
 
-    public void setTimestampUtc(int timestampUtc) {
+    public void setTimestampUtc(Long timestampUtc) {
         this.timestampUtc = timestampUtc;
     }
