+++ backend/src/main/java/org/example/database/repository/ImuRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface ImuRepository extends CrudRepository<Imu, Long> {
     Imu findById(long id);
-    List<Imu> findByTimestampUtcBetween(int start, int end);
+    List<Imu> findByTimestampUtcBetween(Long start, Long end);
 }
